Fixed ICE-5598 (generateRandom) and ICE-5520 (removed openssl/patch.mingw)

+++ cpp/include/IceUtil/Random.h
@@ -16,7 +16,7 @@
 namespace IceUtilInternal
 {
 
-ICE_UTIL_API void generateRandom(char*, int);
+ICE_UTIL_API void generateRandom(char*, size_t);
 ICE_UTIL_API unsigned int random(int = 0);
 
 }
