+++ cpp/src/IceUtil/UUID.cpp
@@ -135,7 +135,7 @@ IceUtil::generateUUID()
     // Randoms by the last 15 bits of the process id.
     //
     char* buffer = reinterpret_cast<char*>(&uuid);
-    IceUtilInternal::generateRandom(buffer, static_cast<int>(sizeof(UUID)));
+    IceUtilInternal::generateRandom(buffer, sizeof(UUID));
 
     //
     // Adjust the bits that say "version 4" UUID
